video-api/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Google APIs
google-generativeai>=0.8.0
google-genai>=0.2.0

# ElevenLabs TTS
elevenlabs>=1.0.0

# Video Processing
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

